requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
Django==3.1.2
django-bootstrap4==2.3.1
django-tinymce==3.1.0
importlib-metadata==2.0.0
Pillow==8.0.0
psycopg2==2.8.6
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.4.1
zipp==3.3.1
